requirements.txt
et-xmlfile==1.1.0
greenlet==1.0.0
mysql==0.0.2
mysqlclient==2.0.3
openpyxl==3.0.7
protobuf==3.15.8
six==1.16.0
SQLAlchemy==1.4.13
